src/dspy-tutorial-5-real-world-use-cases/requirements.txt
dspy-ai==2.6.27
openai==1.88.0
python-dotenv==1.1.0
